+++ tracker/package.json
@@ -6,16 +6,16 @@
   "workspaces": [
     "core/**",
     "plugins/**",
-    "trackers/angular",
-    "trackers/browser",
-    "trackers/react"
+    "trackers/**",
+    "transports/**"
   ],
   "scripts": {
     "clear": "yarn exec make clean && yarn test --clearCache && yarn cache clean",
     "build-core": "yarn workspaces foreach -p --include @objectiv/tracker-core --verbose run build",
     "build-plugins": "yarn workspaces foreach -p --include '@objectiv/plugin**' --verbose run build",
-    "build-trackers": "yarn workspace @objectiv/tracker-browser build && yarn workspace @objectiv/tracker-angular build",
-    "build": "yarn build-core && yarn build-plugins && yarn build-trackers",
+    "build-transports": "yarn workspaces foreach -p --include '@objectiv/transport**' --verbose run build",
+    "build-trackers": "yarn workspace @objectiv/tracker-browser build & yarn workspace @objectiv/tracker-angular build",
+    "build": "yarn build-core && yarn build-plugins & yarn build-transports && yarn build-trackers",
     "prettify": "yarn workspaces foreach --exclude root --verbose run prettify",
     "prettify:generated": "yarn workspace @objectiv/schema prettify:generated && yarn workspace @objectiv/tracker-core prettify:generated",
     "list": "yarn workspaces list",
