+++ tracker/trackers/angular/package.json
@@ -36,7 +36,8 @@
     "npm-publish:verdaccio": "(cd dist && shx cp ../../../verdaccio/.npmrc .npmrc && npm publish && rm -f .npmrc)"
   },
   "peerDependencies": {
-    "@objectiv/tracker-browser": "^0.0.1"
+    "@objectiv/tracker-browser": "^0.0.1",
+    "uuid-random": "^1.3.2"
   },
   "devDependencies": {
     "@angular/compiler": "^12.2.11",
