+++ tracker/trackers/browser/jest.config.js
@@ -11,6 +11,7 @@ module.exports = {
     '@objectiv/plugin-(.*)': '<rootDir>/../../plugins/$1/src',
     '@objectiv/testing-tools': '<rootDir>../../core/testing-tools/src',
     '@objectiv/tracker-core': '<rootDir>/../../core/tracker/src',
+    '@objectiv/transport-(.*)': '<rootDir>/../../transports/$1/src',
   },
   setupFilesAfterEnv: ['jest-extended/all'],
 };
