+++ tracker/trackers/browser/src/index.ts
@@ -104,10 +104,6 @@ export * from './mutationObserver/trackVisibilityVisibleEvent';
 
 export * from './queues/TrackerQueueLocalStorage';
 
-export * from './transports/DebugTransport';
-export * from './transports/FetchAPITransport';
-export * from './transports/XMLHttpRequestTransport';
-
 export * from './BrowserTracker';
 export * from './getOrMakeTracker';
 export * from './getTracker';
