+++ tracker/trackers/react/src/common/factories/makeDefaultTransport.ts
@@ -3,9 +3,9 @@
  */
 
 import { TrackerTransportInterface, TrackerTransportRetry, TrackerTransportSwitch } from '@objectiv/tracker-core';
+import { FetchTransport } from '@objectiv/transport-fetch';
+import { XHRTransport } from '@objectiv/transport-xhr';
 import { ReactTrackerConfig } from '../../ReactTracker';
-import { FetchAPITransport } from '../../transports/FetchAPITransport';
-import { XMLHttpRequestTransport } from '../../transports/XMLHttpRequestTransport';
 
 /**
  * A factory to create the default Transport of Browser Tracker.
@@ -16,8 +16,8 @@ export const makeDefaultTransport = (trackerConfig: ReactTrackerConfig): Tracker
     transport: new TrackerTransportSwitch({
       console: trackerConfig.console,
       transports: [
-        new FetchAPITransport({ endpoint: trackerConfig.endpoint, console: trackerConfig.console }),
-        new XMLHttpRequestTransport({ endpoint: trackerConfig.endpoint, console: trackerConfig.console }),
+        new FetchTransport({ endpoint: trackerConfig.endpoint, console: trackerConfig.console }),
+        new XHRTransport({ endpoint: trackerConfig.endpoint, console: trackerConfig.console }),
       ],
     }),
   });
