+++ tracker/trackers/react/src/index.ts
@@ -10,8 +10,4 @@ export * from './common/factories/makeDefaultTransport';
 
 export * from './queues/TrackerQueueLocalStorage';
 
-export * from './transports/DebugTransport';
-export * from './transports/FetchAPITransport';
-export * from './transports/XMLHttpRequestTransport';
-
 export * from './ReactTracker';
