--- tracker/trackers/react/tests/mocks/localStorageMock.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright 2021 Objectiv B.V.
- */
-
-export const localStorageMock = (function () {
-  let store: any = {};
-  return {
-    getItem: function (key: string) {
-      return store[key];
-    },
-    setItem: function (key: string, value: string) {
-      store[key] = value.toString();
-    },
-    clear: function () {
-      store = {};
-    },
-    removeItem: function (key: string) {
-      delete store[key];
-    },
-  };
-})();
