+++ tracker/transports/debug/README.md
@@ -1,0 +1,29 @@
+# Objectiv DebugTransport
+
+A TrackerTransport that logs incoming events to console.debug
+
+---
+
+## Package installation
+
+```sh
+yarn add @objectiv/transport-debug
+```
+
+### or
+
+```sh
+npm install @objectiv/transport-debug
+```
+
+## Usage
+
+For a detailed usage guide, see the documentation: [https://objectiv.io/docs](https://objectiv.io/docs)
+
+## Copyright and license
+
+Licensed and distributed under the Apache 2.0 License (An OSI Approved License).
+
+Copyright (c) 2021 Objectiv B.V.
+
+All rights reserved.
