+++ tracker/transports/debug/src/DebugTransport.ts
