+++ tracker/transports/debug/src/index.ts
@@ -1,0 +1,5 @@
+/*
+ * Copyright 2021 Objectiv B.V.
+ */
+
+export * from './DebugTransport';
