+++ tracker/transports/fetch/README.md
@@ -1,0 +1,29 @@
+# Objectiv FetchTransport
+
+A TrackerTransport based on Fetch API
+
+---
+
+## Package installation
+
+```sh
+yarn add @objectiv/transport-fetch
+```
+
+### or
+
+```sh
+npm install @objectiv/transport-fetch
+```
+
+## Usage
+
+For a detailed usage guide, see the documentation: [https://objectiv.io/docs](https://objectiv.io/docs)
+
+## Copyright and license
+
+Licensed and distributed under the Apache 2.0 License (An OSI Approved License).
+
+Copyright (c) 2021 Objectiv B.V.
+
+All rights reserved.
