+++ tracker/transports/fetch/jest.config.js
@@ -6,7 +6,5 @@ module.exports = {
   moduleNameMapper: {
     '@objectiv/testing-tools': '<rootDir>../../core/testing-tools/src',
     '@objectiv/tracker-core': '<rootDir>/../../core/tracker/src',
-    '@objectiv/tracker-core-react': '<rootDir>/../../core/react/src',
-    '@objectiv/transport-(.*)': '<rootDir>/../../transports/$1/src',
   },
 };
