+++ tracker/transports/fetch/src/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2021 Objectiv B.V.
+ */
+
+export * from './defaultFetchFunction';
+export * from './defaultFetchOptions';
+export * from './FetchTransport';
