+++ tracker/transports/xhr/.prettierignore
@@ -1,0 +1,29 @@
+# Logs
+*.log
+
+# Yarn 2
+.yarn/*
+!.yarn/releases
+!.yarn/plugins
+
+# Dependencies
+node_modules/
+
+# OS X temporary files
+.DS_Store
+
+# Test coverage reports
+coverage/
+
+# Distribution files
+dist/
+
+# IDE files
+.idea/
+.vscode/
+
+# Markdown
+*.md
+
+# React Tracker Live testing App
+tests/live
