+++ tracker/transports/xhr/README.md
@@ -1,0 +1,29 @@
+# Objectiv XHRTransport
+
+A TrackerTransport based on XMLHttpRequest API
+
+---
+
+## Package installation
+
+```sh
+yarn add @objectiv/transport-xhr
+```
+
+### or
+
+```sh
+npm install @objectiv/transport-xhr
+```
+
+## Usage
+
+For a detailed usage guide, see the documentation: [https://objectiv.io/docs](https://objectiv.io/docs)
+
+## Copyright and license
+
+Licensed and distributed under the Apache 2.0 License (An OSI Approved License).
+
+Copyright (c) 2021 Objectiv B.V.
+
+All rights reserved.
