+++ tracker/transports/xhr/src/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Copyright 2021 Objectiv B.V.
+ */
+
+export * from './defaultXHRFunction';
+export * from './XHRTransport';
